+++ assets/static/project-map/stores/index.js
@@ -12,9 +12,20 @@ export {
   viewportSites,
   tagFilteredSites,
   sidebarSites,
+  syncTagFilterVisibility,
+  isSiteMapVisible,
+  toggleSiteManualHidden,
   annotateImportPoints,
   bulkTagInitialCounts,
   computeBulkTagOps,
 } from './sites.js'
+export {
+  isPeakMapVisible,
+  visiblePeaksForMap,
+  sortedPeaksList,
+  viewportPeaks,
+  sidebarPeaks,
+  hiddenPeakSlugs,
+} from './peaks.js'
 export { setLinksPayload, bumpWarmPriorities } from './links.js'
 export { setSimulation } from './simulation.js'